parking-cloud/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
redis
orjson
pytz
python-multipart
qrcode[pil]
pillow
